Extract descendence from Node
