+++ lib/rubocop/ast.rb
@@ -7,6 +7,7 @@ require 'set'
 require_relative 'ast/ext/range'
 require_relative 'ast/ext/set'
 require_relative 'ast/node_pattern'
+require_relative 'ast/node/mixin/descendence'
 require_relative 'ast/sexp'
 require_relative 'ast/node'
 require_relative 'ast/node/mixin/method_identifier_predicates'
